requirements.txt
pyglet==2.0.17
moderngl==5.8.2
numpy==1.26.4
PyGLM==2.7.1
Pillow==10.4.0
noise==1.2.2
requests==2.32.3